Added initial search results section, two columns
-Added an intial search results section that displays igt ids and
additional information depending on what tier the user searched in.

If the user searched the phrases tier than the phrase that is stored in
the phrase tier for each igt, is displayed next to the igt ids.

If the user searched the word tier, than the words that make up the
phrases described above are displayed next to each igt id.

If the user searched the morphemes tier than the morphemes that make up
the words described above are displayed next to the igt ids.

-Once one of the igts is selected in the initial search result section,
the user is taken to a results section with two columns. On the left
there is a list of igt ids for found igts. On the right is the data for
the currently selected igt. The igts on the left can be selected.

-Added a button to allow the user to go back to the intial search
results section.

-Added styling to the intial search results section and the button that
allows the user to go back to it.