Added row for each match in the intial results
When there is multiple matches in one igt, one match is centered and
highlighted while the rest are given a red border. Each match gets a
line where it is centered and highlighted.

If a row in the intial search result section is hovered over, the full
contents of the table cell is displayed.